requirements.txt
Django
django-cors-headers
djangorestframework
markdown
django-filter
psycopg2